align with SWA-85
