+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/ContactInformation/ContactInformation.tests.tsx 
@@ -1,0 +1,1 @@
+// this is a test file
